Moved form input to components folder
