+++ README.md
@@ -1,3 +1,9 @@
+In case got error on running tauri
+```
 sudo xcode-select --reset
 sudo xcodebuild -license
+```
+To open console at dev mode:
+```
 cmd + option + i
+```
